HF DAAscore to switch to khashv2 (12 hours after testnet launch):
We've experimented with 3 different hashing algorithm in testnet,
named khashv1, khashv1.5 and khashv2.

khashv1   = KarlsenHashv1
khashv1.5 = FishHash
khashv2   = FishHash Plus (KarlsenHashv2)

However at the end we've decided to use FishHash Plus as it was
result of security review of the FishHash algorithm. The PoW code
for FishHash is still there, but the expected blocks are either
version 1 or version 2. Therefore we decided to restart testnet-1
with shorter HF DAA score and don't need to maintain code for
special case in the network.